### 2016-08-16 #71 @sckott
rdpla package
Client for the Digital Public Library of America (DPLA), which holds metadata from lots of museums, focusing on the US.

```
Package: rdpla
Title: Client for the Digital Public Library of America ('DPLA')
Description: Interact with the Digital Public Library of America
    ('DPLA') 'REST' 'API' from R, including search and more.
Version: 0.0.7.9300
Authors@R: c(person("Scott", "Chamberlain", role = c("aut", "cre"),
    email = "myrmecocystus@gmail.com"))
LazyData: yes
    jsonlite (>= 1.0),
    data.table
    roxygen2 (>= 5.0.1),
    scales
```

https://github.com/ropensci/rdpla

Academics that can use the data DPLA has for their research on a variety of topics (e..g, a certain type of photograph through time, or works by a certain artist, or a spatial analysis of works), or collection managers that submit metadata to DPLA so they can analyze data on their own collections.

__No__

- [x] Do you intend for this package to go on CRAN?  __YES__
- [ ] Do you wish to automatically submit to the [Journal of Open Source Software] (http://joss.theoj.org/)? If so:   __no__
    - [ ] The package contains a [`paper.md`](http://joss.theoj.org/about#paper_structure) with a high-level description.

Note that Travis build is failing right now, but b/c of some key server error - not fault of the pkg pretty sure, so should be working again soon. Passes check fine locally.
