### 2016-08-26 #71 @noamross
### Editor checks:

- [x] **Fit**: The package meets or criteria [fit](policies.md#fit) and [overlap](policies.md#fit)

---

#### Editor comments

Will be seeking reviewers, but in the meantime:

There was a typo in the README in the appropriate API key variable name that precluded local testing.  I sent a PR.

Automated review tools gave these results (I recommend both authors and reviewers examine these more closely):

`goodpractice::gp()`:

```
  ✖ write unit tests for all functions, and all package code in general. At
    least some lines are not covered in this package.

    R/collections.r:34:NA
    R/collections.r:35:NA
    R/collections.r:36:NA
    R/collections.r:37:NA
    R/collections.r:38:NA
    ... and 195 more lines

  ✖ avoid long code lines, it is bad for readability. Also, many people prefer
    editor windows that are about 80 characters wide. Try make your lines
    shorter than 80 characters

    inst/ignore/dpla_items_old.r:76:1
    inst/ignore/dpla_items_old.r:76:1
    inst/ignore/dpla_items_old.r:76:1
    inst/ignore/dpla_items_old.r:76:1
    inst/ignore/dpla_items_old.r:91:1
    ... and 15 more lines

  ✖ avoid sapply(), it is not type safe. It might return a vector, or a list,
    depending on the input data. Consider using vapply() instead.

    inst/ignore/dpla_items_old.r:87:21
    inst/ignore/dpla_plot.r:45:29
    R/collections.r:46:12
    R/items.r:220:11
    R/items.r:326:11

  ✖ avoid 1:length(...), 1:nrow(...), 1:ncol(...), 1:NROW(...) and 1:NCOL(...)
    expressions. They are error prone and result 1:0 if the expression on the
    right hand side is zero. Use seq_len() or seq_along() instead.

    inst/ignore/dpla_plot.r:27:21
```

While test are running locally, `covr::package_coverage()` yields 0% test coverage.  I suggest you look into why.
---

Reviewers:
Due date:
