### 2016-09-15 #71 @jsonbecker
- [X] A statement of need clearly stating problems the software is designed to solve and its target audience in README
- [X] Installation instructions: for the development version of package and any non-standard dependencies in README
- [X] Vignette(s) demonstrating major functionality that runs successfully locally

I think it would be helpful here to describe a bit more about the API. For example, the main way the data are accessed appears to be through `dpla_items` that are items in `dpla_collections`. I think discussing this up front, their contents, and how they interact would provide useful information to new users in addition to the solid examples provided.

- [X] Function Documentation: for all exported functions in R help
- [X] Examples for all exported functions in R Help that run successfully locally
- [ ] Community guidelines including contribution guidelines in the README or CONTRIBUTING, and URL, Maintainer and BugReports fields in DESCRIPTION

No Maintainer in DESCRIPTION.
##  Installation:
- [X] Installation succeeds as documented.
- [ ] Functionality: Any functional claims of the software been confirmed.
  - [ ] Using the example for `dpla_items`, the following warning is issued:

```r
In if (class(output2) %in% "factor") { :
  the condition has length > 1 and only the first element will be used
```
I think this is due to the code [here](https://github.com/ropensci/rdpla/blob/9728ea20378fa0c10be3d56f690e047525a08763/R/items.r#L206). This check is being done to `output2` which has more than one class due to the use of `dplyr` and `tibble`, e.g. `class(output2)` results in `"tbl_df"     "tbl"        "data.frame"`. To reproduce, use `dpla_items(q="fruit", fields=c("id","publisher","format"), ...)`.

In order to address this, you can do `if (any(class(output2) %in% 'factor')`.

[X] - Performance: Any performance claims of the software been confirmed.
[X] - Automated tests: Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.

I do think there can be some improvement here. Not all cases are covered for an API with a lot of options. One thing that could be tested without a `skip_on_cran` is that the API requests are properly formed without actually submitting them. This way you can try one call with all parameters and see that the GET request is properly formatted as expected. I don't have particularly great advise on how to run this test authentically from the `dpla_items` or `dpla_collections` calls, but I am relatively certain that I've seen this style of test in other packages. Maybe there's a return condition that can be applied only if running a test.

[X] - Packaging guidelines: The package conforms to the rOpenSci packaging guidelines

## Final approval (post-review)

[ ] - The author has responded to my review and made changes to my satisfaction. I recommend approving this package.
