### 2016-09-18 #71 @fmichonneau
This is a very nice and concise package that does what it says it does. Not
knowing the existence of this resource before reviewing this package (and
probably also because it's not my field of expertise), I wasn't sure how this
package could be used until I saw the use cases described in #3. Maybe
implementing the church visualization example would have helped me earlier to
understand how the package could be used.

Tests and checks pass on my system. `goodpractice` detected the same things as
Noam noted.

```
> session_info("rdpla")
Session info ------------------------------------------------------------------
 system   x86_64, linux-gnu
 ui       X11
 language en_US:en
 collate  en_US.UTF-8
 tz       posixrules

Packages ----------------------------------------------------------------------
 package    * version    date       source
 assertthat   0.1        2013-12-06 CRAN (R 3.3.0)
 chron        2.3-47     2015-06-24 CRAN (R 3.3.0)
 curl         1.2        2016-08-13 CRAN (R 3.3.1)
 data.table   1.9.6      2015-09-19 CRAN (R 3.3.0)
 httr         1.2.1      2016-07-03 CRAN (R 3.3.1)
 jsonlite     1.1        2016-09-14 CRAN (R 3.3.1)
 lazyeval     0.2.0      2016-06-12 CRAN (R 3.3.0)
 mime         0.5        2016-07-07 CRAN (R 3.3.1)
 openssl      0.9.4      2016-05-25 CRAN (R 3.3.0)
 R6           2.1.3      2016-08-19 CRAN (R 3.3.1)
 Rcpp         0.12.7     2016-09-05 CRAN (R 3.3.1)
 rdpla      * 0.0.7.9500 2016-09-18 local (fmichonneau/rdpla@NA)
 tibble       1.2        2016-08-26 CRAN (R 3.3.1)
```

- [x] **A statement of need** clearly stating problems the software is designed
  to solve and its target audience in README

  Not knowing about this resource, I found that the README (and vignette) could
  include a little more information about the kind of information and documents
  the library holds, and information about what constitute a "collection" or an
  "item".

- [x] **Installation instructions:** for the development version of package and
  any non-standard dependencies in README

  No problem with the installation of the package, and getting the API
  key. If the vignette includes the installation instructions, it might be worth
  also adding how to obtain an API key there.

- [x] **Vignette(s)** demonstrating major functionality that runs successfully
  locally

  The vignette illustrates many of the functionalities of the package. Several
  examples were producing warnings. I think that the pull request #20 should fix
  them. Unless I missed something, it might also be worth noting that the
  examples and the vignette can only run locally if the API key is included in
  the `.Rprofile`. Checks on Travis-CI seem to currently be failing because of
  that (at least for the PR I submitted).

  There are help files for the 3 exported functions.

  * For `dpla_items()`, under the `Value` header, only the contents of `meta`
    and `data` are listed, and not `facets`.

  * For `dpla_collections()`, in the list of arguments, for the parameter
    `sort_by`, a sentence is incomplete ("is all fields."), and indicates to see
    the "Details" section that is not present.

  * For `get_key()`, you may want to remove your email address (if it's a real
    one) to avoid getting a notification every time someone run
    `example(get_key)

  There are also help pages for `country_codes` and `language_codes` but neither
   exist as functions (i.e., defined as `NULL` in `rdpla-package.R` file).

- [x] **Packaging guidelines**: The package conforms to the rOpenSci packaging
  guidelines

  You may want to consider adding a NEWS file and integration for code coverage with `covr`.

Estimated hours spent reviewing: 2.5 hours
