### 2016-09-29 #71 @sckott
response to @jsonbecker review

> I think it would be helpful here to describe a bit more about the API. For example, the main way the data are accessed appears to be through dpla_items that are items in dpla_collections. I think discussing this up front, their contents, and how they interact would provide useful information to new users in addition to the solid examples provided.

Added more discussion of items vs. collections in the readme and vignette.

> No Maintainer in DESCRIPTION.

The `Authors@R` syntax is in place of that.

> Warning message in `if (class(output2) %in% "factor") { : ...`

this was fixed by a PR https://github.com/ropensci/rdpla/pull/20 - thanks @fmichonneau  !

> Tests: I do think there can be some improvement here. Not all cases are covered for an API with a lot of options. One thing that could be tested without a skip_on_cran is that the API requests are properly formed without actually submitting them. This way you can try one call with all parameters and see that the GET request is properly formatted as expected. I don't have particularly great advise on how to run this test authentically from the dpla_items or dpla_collections calls, but I am relatively certain that I've seen this style of test in other packages. Maybe there's a return condition that can be applied only if running a test.

I've re-worked the two functions to separate them into 2 parts for each of the two main exported functions:

* an unexported internal fxn that does query preparation and http request
* the exported fxn does further parsing and cleanup

And added tests for the internal fxns. Which means that HTTP requests are being tested twice, but only the exported fxn tests additionally test the parsing to data.frame, etc.

-------------

response to @fmichonneau  review:

> I wasn't sure how this package could be used until I saw the use cases described in #3

good point, i'll try to implement that as a vignette

> Not knowing about this resource, I found that the README (and vignette) could
include a little more information about the kind of information and documents
the library holds, and information about what constitute a "collection" or an
"item".

Thanks. I added more info on what DPLA has, and diff. btw collections and items

> If the vignette includes the installation instructions, it might be worth
also adding how to obtain an API key there.

done, thanks!

> it might also be worth noting that the examples and the vignette can only run locally if the API key is included in the .Rprofile.

added note to readme, thanks

> For dpla_items(), under the Value header, only the contents of meta and data are listed, and not facets.

fixed, thanks!

> For dpla_collections(), in the list of arguments, for the parameter sort_by, a sentence is incomplete ("is all fields."), and indicates to see the "Details" section that is not present.

fixed, thanks!

> For get_key(), you may want to remove your email address (if it's a real one) to avoid getting a notification every time someone run `example(get_key)

done, thanks!

> There are also help pages for country_codes and language_codes but neither
exist as functions (i.e., defined as NULL in rdpla-package.R file).

`country_codes` and `language_codes` are datasets (another is `dpla_fields`), so the documentation has NULL at bottom. I thought this was standard for documenting datasets. Is something else better?

> You may want to consider adding a NEWS file and integration for code coverage with covr.

thanks, adding `NEWS.md` file. covr added
