### 2016-09-29 #71 @sckott
Now vignette added, https://github.com/ropensci/rdpla/blob/master/vignettes/use_case_churches.Rmd
