### 2016-09-30 #71 @noamross
Thanks @sckott! @jsonbecker and @fmichonneau, please review changes and let us know if they address your comments.  If they do, let us know and check off any remaining boxes in your review. Let us know of any outstanding issues.
