### 2016-10-02 #71 @jsonbecker
These improvements look good to me. 👍
