### 2016-10-07 #71 @fmichonneau
Thanks @sckott! It all looks good to me. I'd just suggest you add a little bit of documentation for the datasets. I think it's best when there is at least some info about them. You can check the documentation for `iris` or `diamonds` for some ideas. There are some details in Hadley's book: http://r-pkgs.had.co.nz/data.html

You may also want to update the README to reflect that the vignette demonstrating the use case it available.

Nice work!
