### 2016-10-07 #71 @sckott
thanks @jsonbecker and @fmichonneau  for your reviews!  noted about documenting the datasets and the vignette, will do
