### 2016-10-07 #71 @sckott
okay, updated docs for datasets and mentioned the vignettes., will link to them once it's on cran
