### 2016-10-10 #71 @noamross
Thanks, all.  @sckott, two small things from a last check:

-   From `goodpractice::gp()` I still see several in-function `sapply()` calls, and a `1:nrow()`
-   Nice new vignette format!  Question - will these be suitable for CRAN, or will we also need `html_vignette()`-formatted versions?
