### 2016-10-11 #71 @sckott
* will fix
* I'm pretty sure that vignette format is fine, just that it's a bit larger file size than the vignette specific format
