### 2016-10-11 #71 @sckott
@noamross  the `sapply` usage has been changed. the `1:nrow` is in an old file in `inst/ignore`, so is not used in the package.
