### 2016-10-24 #71 @sckott
@noamross should we close now?
